нашли ошибочки и исправили ошибочки

+++ AbstractUniversity/AbstractUniversityImplementation/Migrations/20200604190117_InitialCreate.cs
@@ -31,7 +31,7 @@ namespace AbstractUniversityImplementation.Migrations
                     Id = table.Column<int>(nullable: false)
                         .Annotation("SqlServer:Identity", "1, 1"),
                     DisciplineName = table.Column<string>(nullable: false),
-                    Price = table.Column<int>(nullable: false)
+                    Price = table.Column<decimal>(nullable: false)
                 },
                 constraints: table =>
                 {
